Add TokenInformation struct
