+++ ios/Anywhere Reader/Anywhere Reader/Services/APIService.swift
@@ -46,19 +46,3 @@ class APIService {
         }.resume()
     }
 }
-
-struct TokenInformation: Codable {
-    let accessToken: String
-    let expiresIn: Int
-    let tokenType: String
-    let scope: String
-    let refreshToken: String
-    
-    enum CodingKeys: String, CodingKey {
-        case accessToken = "access_token"
-        case expiresIn = "expires_in"
-        case tokenType = "token_type"
-        case scope = "scope"
-        case refreshToken = "refresh_token"
-    }
-}
